+++ examples/heroku-button/nitro.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from 'nitro';
+
+export default defineConfig({
+  modules: ['workflow/nitro'],
+  plugins: ['./server/plugins/start-world.ts'],
+  serverDir: './server',
+});