+++ examples/heroku-button/server/plugins/start-world.ts
@@ -1,0 +1,11 @@
+import { definePlugin } from 'nitro';
+
+export default definePlugin(async (nitro) => {
+  const { getWorld } = await import('workflow/runtime');
+  const world = getWorld();
+  await world.start?.();
+
+  nitro.hooks.hook('close', async () => {
+    await world.close?.();
+  });
+});